code refactoring
header / source file division,
fix korean language not print in windows cmd.
add / rm function wrtie
