+++ Memo.cpp
@@ -1,0 +1,10 @@
+#include "header\Memo.h"	
+
+Memo::Memo(const string name)
+	: name(name),
+	tagindex(0)
+{
+}
+
+Memo::~Memo() {
+}
